requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
pdfkit
jinja2
PyMuPDF

# fastapi==0.111.0
# uvicorn==0.29.0
# groq==0.8.0
# python-dotenv==1.0.1
# Jinja2==3.1.4
# WeasyPrint==62.3